Refractor Database
Refractored the database.py file with functional and structural changes.
Removed unnessarcary lines to docstrings.
Restructured many of the docstrings format to compact the comments.
Reworded many docstrings.
Added the new commit and rollback class methods to allow controlled
committing
and rollbacks using the database object.
Changed the purpose of the commit method, previously the commit method
handled the execution of SQL statements such as the insert statements that
are immediantly committed. Now all queries are handled by the query method
and commit only commits the queries.
Added automatic commit and rollback functionality to the __exit__ magic
method, if an error occurs within a with context block it is automatically
rolled back, otherwise, it is automatically commited.
Added a __repr__ magic method to allow for easier debugging.
Removed the iterate method as it's method of iteration was dated and
unpythonic and the iterate method has not been used in this project.
Added the limit keyword to the query method as postgres does not support
the LIMIT keyword this allows limiting.
Allowed the query method to handle queries which do not return any data
which previously raised a ProgrammingError.
Allowed the exists method to handle empty with parameters to allow
checking if a table exists at all.
Ensured that booleans are actually returns from exists rather than falsey
evaulation.
Added the __bool__ magic method to check if the connection to the database
has been closed.
Added the __del__ magic method to allow for pythonic closing of the
database as well as proper disposal of the object.