Restructured Assessment Index
Performed a major restructuring of the assessment index to combine the
code of assessment and group items as a lot of the code is similar,
instead it just alters the code slightly based on which type the item is.

Added a modal for updating the result of an assessment.

Added javascript code to perform REST actions.

Altered the assessment creation api to a more modern url and return type.
Also included error checking.

Updated add_assessment template to use javascript to add assessment rather
than HTML forms.